requirements.txt
fastapi==0.99.1
openai==0.27.10
pydantic==1.10.13
uvicorn==0.25.0
langchain==0.0.349
jinja2==3.1.2
websockets==12.0
python-dotenv==1.0.0
llama-cpp-python==0.2.27
huggingface-cli